requirements.txt
absl-py==2.1.0
attrs==24.2.0
cffi==1.17.1
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==24.3.25
fonttools==4.55.0
jax==0.4.35
jaxlib==0.4.35
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.18
ml_dtypes==0.5.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.1.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2
